#21629: Restructure yolo eval folder
